requirements.txt
Django==5.0.3
djangorestframework==3.14.0
python-dotenv==1.0.1
drf-spectacular==0.27.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2==2.9.9